app/backend/requirements.txt
smolagents==1.7.0
boto3==1.35.7
fastapi==0.112.2
python-dotenv==1.0.1
uvicorn==0.30.6
openai>=1.55.3
smolagents[litellm]   